+++ lib/td_ai/application.ex
@@ -35,7 +35,8 @@ defmodule TdAi.Application do
     [
       TdAi.Milvus,
       TdAi.NxServings,
-      TdAi.GenAi
+      TdAi.GenAi,
+      TdAi.Embeddings.Server
     ]
   end