+++ priv/repo/migrations/20250411092012_add_enabled_at_to_indices.exs
@@ -1,0 +1,9 @@
+defmodule TdAi.Repo.Migrations.AddEnabledAtToIndices do
+  use Ecto.Migration
+
+  def change do
+    alter table(:indices) do
+      add :enabled_at, :utc_datetime_usec
+    end
+  end
+end