+++ priv/repo/migrations/20250411092408_add_unique_index_to_collection_name_in_indices.exs
@@ -1,0 +1,7 @@
+defmodule TdAi.Repo.Migrations.AddUniqueIndexToCollectionNameInInices do
+  use Ecto.Migration
+
+  def change do
+    create unique_index(:indices, [:collection_name])
+  end
+end